requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.6.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.330
langcodes==3.3.0
langsmith==0.0.57
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mido==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==2.1.2
pathtools==0.1.2
Pillow==10.1.0
preshed==3.0.9
prettytable==3.9.0
protobuf==4.25.0
psutil==5.9.6
pyarrow==14.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
streamlit==1.28.1
tenacity==8.2.3
termcolor==2.3.0
textstat==0.7.3
thinc==8.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wandb==0.15.12
wasabi==1.1.2
wcwidth==0.2.9
weasel==0.3.3
yarl==1.9.2
zipp==3.17.0
